+++ src/yaplox/stmt.py
@@ -31,6 +31,10 @@ class StmtVisitor(ABC):
     def visit_var_stmt(self, stmt: "Var"):
         raise NotImplementedError
 
+    @abstractmethod
+    def visit_while_stmt(self, stmt: "While"):
+        raise NotImplementedError
+
 
 class Stmt(ABC):
     @abstractmethod
@@ -84,3 +88,13 @@ class Var(Stmt):
     def accept(self, visitor: StmtVisitor):
         """ Create a accept method that calls the visitor. """
         return visitor.visit_var_stmt(self)
+
+
+class While(Stmt):
+    def __init__(self, condition: Expr, body: Stmt):
+        self.condition = condition
+        self.body = body
+
+    def accept(self, visitor: StmtVisitor):
+        """ Create a accept method that calls the visitor. """
+        return visitor.visit_while_stmt(self)